+++ finder-template-generator-ssg/docs/changelog.md
@@ -9,6 +9,7 @@ _The whole finder CHANGELOG_
 - added cache viewer
 - Moved every site on the finder website to TSX instead
   of TS (template viewer) - 06.09.2026
+- add code to the ssg plugin which runs on build time - 07.09.2026
 
 ## 0.3.15 - 02.09.2025